Changing the order and adding context in #Set up
In the Scenario section, we say that we're going to crack passwords and then set up our own uncrackable ones.  In the next section, we say we are going to use two tools: keepassx and john the ripper.    If someone is not aware of password managers, they likely won't know that keepassx is one.  So it makes the setup section a little bit confusing, cos it goes straight into installing stuff but not explaining why you're installing it.