Moving the additional note about hashes
With #Sanity Check, you are asking the reader to grab two files and use this to learn the basics about password cracking.  The first side note within this section, though, does not relate directly to this and is instead about how you generated a hashed string.  This is an interesting note, but made the section confusing overall.  For me, I read this sidenote as a "here's a tip for doing this sanity-check", and thus that I was supposed to be making the sanity-check file.  I didn't know to read this as a "behind-the scenes cool note".

So I moved the note itself down to further discussion, under the hash subsection.  That may not be the right place subsection for it, but I think overall that it belongs under further discussion.