Softened wording on line 360
If this article is targeted towards non-tech-industry folks, I would avoid wording like "By now it should be obvious" as that is telling the reader they should be able to figure this out, and if they didn't or are still confused, they are alone in this.  I propose changing it to "The reason is simple" as that puts no expectation on the reader.  

I might be overthinking this though!